refactoring package names to support Java 9 modularization
