+++ jgnash-fx/src/main/java/jgnash/uifx/control/DetailedDecimalTextField.java
@@ -29,7 +29,7 @@ import javafx.scene.layout.ColumnConstraints;
 import javafx.scene.layout.GridPane;
 import javafx.scene.layout.Priority;
 import javafx.scene.layout.Region;
-import jgnash.resource.font.FontAwesomeLabel;
+import jgnash.uifx.resource.font.FontAwesomeLabel;
 import jgnash.util.NotNull;
 
 import java.math.BigDecimal;
