+++ jgnash-fx/src/main/java/jgnash/uifx/resource/cursor/CustomCursor.java
@@ -15,7 +15,7 @@
  *  You should have received a copy of the GNU General Public License
  *  along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package jgnash.resource.cursor;
+package jgnash.uifx.resource.cursor;
 
 import javafx.scene.Cursor;
 import javafx.scene.ImageCursor;
