+++ jgnash-fx/src/main/resources/jgnash/uifx/control/wizard/WizardDialog.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.StackPane?>
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 <GridPane maxHeight="Infinity" maxWidth="Infinity" prefHeight="480.0" prefWidth="750.0"
           xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8"
           fx:controller="jgnash.uifx.control.wizard.WizardDialogController" styleClass="form, dialog">
