+++ jgnash-fx/src/main/resources/jgnash/uifx/dialog/ChangePasswordDialog.fxml
@@ -12,7 +12,7 @@
 <?import javafx.scene.control.PasswordField?>
 <?import javafx.scene.control.TitledPane?>
 
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 <BorderPane fx:controller="jgnash.uifx.dialog.ChangeDatabasePasswordDialogController"
             xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" styleClass="dialog">
     <center>
