+++ jgnash-fx/src/main/resources/jgnash/uifx/dialog/ImportScriptsDialog.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 
 <GridPane xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8" minHeight="-Infinity"
           fx:controller="jgnash.uifx.dialog.ImportScriptsDialogController" styleClass="dialog, form">
