+++ jgnash-fx/src/main/resources/jgnash/uifx/dialog/options/TransactionNumberDialog.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 
 <GridPane fx:controller="jgnash.uifx.dialog.options.TransactionNumberDialogController" minHeight="-Infinity"
           xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" styleClass="form, dialog">
