+++ jgnash-fx/src/main/resources/jgnash/uifx/report/jasper/JasperViewerDialog.fxml
@@ -13,7 +13,7 @@
 <?import javafx.scene.layout.BorderPane?>
 <?import javafx.scene.layout.StackPane?>
 <?import javafx.scene.layout.VBox?>
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 
 <StackPane fx:id="stackPane" fx:controller="jgnash.uifx.report.jasper.JasperViewerDialogController"
            xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1">
