+++ jgnash-fx/src/main/resources/jgnash/uifx/views/register/InvestmentRegisterTable.fxml
@@ -3,7 +3,7 @@
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
 
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 <BorderPane xmlns="http://javafx.com/javafx/8" xmlns:fx="http://javafx.com/fxml/1" fx:controller="jgnash.uifx.views.register.InvestmentRegisterTableController">
    <top>
        <TitledPane focusTraversable="false" expanded="false">
