+++ jgnash-fx/src/main/resources/jgnash/uifx/wizard/file/NewFileOne.fxml
@@ -7,7 +7,7 @@
 <?import javafx.scene.layout.ColumnConstraints?>
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.RowConstraints?>
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 <?import jgnash.uifx.control.DataStoreTypeComboBox?>
 
 <?import javafx.scene.control.Tooltip?>
