+++ jgnash-fx/src/main/resources/jgnash/uifx/wizard/file/NewFileThree.fxml
@@ -8,7 +8,7 @@
 <?import javafx.scene.layout.GridPane?>
 <?import javafx.scene.layout.RowConstraints?>
 <?import javafx.scene.layout.VBox?>
-<?import jgnash.resource.font.FontAwesomeLabel?>
+<?import jgnash.uifx.resource.font.FontAwesomeLabel?>
 
 <GridPane xmlns:fx="http://javafx.com/fxml/1" xmlns="http://javafx.com/javafx/8" maxWidth="Infinity"
           fx:controller="jgnash.uifx.wizard.file.NewFileThreeController" styleClass="form">
